Correct use of iat and exp?
According to the spec, iat (issued at time) identifies the time at which the
JWT was issued while exp "identifies the expiration time on or after which the
token MUST NOT be accepted for processing". The specs in test_iam_token_manager
appear to reverse these? I could be wrong but putting in here to set up for the
next test.

Signed-off-by: Moses Mendoza <mendoza.moses@gmail.com>